Move secret errors to seperate file
